requirements.txt
certifi==2022.12.7
cgcomponents==2.0.2
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
loguru==0.6.0
mysql-connector-python==8.0.32
protobuf==3.20.3
pyodbc==4.0.35
python-dotenv==1.0.0
requests==2.28.2
Unidecode==1.3.6
urllib3==1.26.15
win32-setctime==1.1.0
